+++ digestible/src/digestible/std_types.rs
@@ -1,7 +1,7 @@
 use crate::digestible::internal_macros::impl_for_hashable_hack;
 use crate::{DigestWriter, Digestible};
 use byteorder::ByteOrder;
-use std::collections::{BTreeMap, HashMap, HashSet};
+use std::collections::{HashMap, HashSet};
 use std::ffi::{OsStr, OsString};
 use std::path::{Path, PathBuf};
 
@@ -13,7 +13,7 @@ impl_for_hashable_hack!(Path);
 
 impl<K: Digestible, V: Digestible> Digestible for HashMap<K, V> {
     fn digest<B: ByteOrder, W: DigestWriter>(&self, writer: &mut W) {
-        digest_iter::<(&K, &V), B, W, _>(self.iter(), writer);
+        digest_iter::<_, B, W, _>(self.iter(), writer);
     }
 
     fn digest_native<W: DigestWriter>(&self, writer: &mut W) {
@@ -21,14 +21,6 @@ impl<K: Digestible, V: Digestible> Digestible for HashMap<K, V> {
     }
 }
 
-impl<K: Digestible, V: Digestible> Digestible for BTreeMap<K, V> {
-    fn digest<B: ByteOrder, W: DigestWriter>(&self, writer: &mut W) {
-        digest_iter::<(&K, &V), B, W, _>(self.iter(), writer);
-    }
-    fn digest_native<W: DigestWriter>(&self, writer: &mut W) {
-        digest_native_iter::<_, W, _>(self.iter(), writer);
-    }
-}
 impl<V: Digestible> Digestible for HashSet<V> {
     fn digest<B: ByteOrder, W: DigestWriter>(&self, writer: &mut W) {
         digest_iter::<_, B, W, _>(self.iter(), writer);